use core.subset
