+++ clisops/__init__.py
@@ -8,7 +8,3 @@ __contact__ = "eleanor.smith@stfc.ac.uk"
 __copyright__ = "Copyright 2018 United Kingdom Research and Innovation"
 __license__ = "BSD"
 __version__ = "0.1.0"
-
-from .utils import *
-from .ops import *
-from .config import *
