--- clisops/ops/regrid.py
@@ -1,5 +1,0 @@
-__author__ = """Elle Smith"""
-__contact__ = "eleanor.smith@stfc.ac.uk"
-__copyright__ = "Copyright 2018 United Kingdom Research and Innovation"
-__license__ = "BSD"
-__version__ = "0.1.0"
