+++ clisops/utils/__init__.py
@@ -1,9 +1,3 @@
-__author__ = """Elle Smith"""
-__contact__ = "eleanor.smith@stfc.ac.uk"
-__copyright__ = "Copyright 2018 United Kingdom Research and Innovation"
-__license__ = "BSD"
-__version__ = "0.1.0"
-
 from .get_coords import *
 from .get_data import *
 from .map_args import *
