+++ clisops/utils/get_coords.py
@@ -1,10 +1,3 @@
-__author__ = """Elle Smith"""
-__contact__ = "eleanor.smith@stfc.ac.uk"
-__copyright__ = "Copyright 2018 United Kingdom Research and Innovation"
-__license__ = "BSD"
-__version__ = "0.1.0"
-
-
 def get_coord_by_attr(dset, attr, value):
     coords = dset.coords
