+++ clisops/utils/get_data.py
@@ -1,9 +1,3 @@
-__author__ = """Elle Smith"""
-__contact__ = "eleanor.smith@stfc.ac.uk"
-__copyright__ = "Copyright 2018 United Kingdom Research and Innovation"
-__license__ = "BSD"
-__version__ = "0.1.0"
-
 from clisops.utils.get_coords import get_latitude, get_longitude
